Add dbus emulation of OHM API
Add a dbus-activated daemon which implements a small part of the OHM API.
This will allow the old ways of freezing/unfreezing the dcon to continue
to work (currently used by sugar and olpc-utils).

In the future, more of the OHM API could be emulated, and a powerd-specific
API could be added too.