Inhibit suspend during wifi scanning/connecting
powerd-dbus is now launched by powerd and monitors NetworkManager wifi
devices. When those devices are establishing connections, it tells
powerd not to suspend. NetworkManager v0.7 and v0.8 are both supported.

Support is also added to monitor wpa_supplicant's scanning signals,
but this was found to be too hard to do on 0.6 (incomplete dbus API),
and while I made it work for v0.7, it requires an upstream patch
which is not yet merged.