arrange for sending the switch device pathnames to powerd
powerd needs the device name to do state checking on the
lid and ebook switches.  since olpc-switchd has the name,
and it's somewhat important that powerd use the same device,
we pass it along.  it gets sent on startup (in case olpc-switchd
starts after powerd), on SIGUSR1 (in case it starts before), and
for good measure, later on with any close/open events.