first cut at using /sys/power/wakeup_count
after we've read /sys/power/wakeup_count, writing the same value
back will only succeed if no new calls to pm_wakeup_event() have
been made in the kernel.  if it doesn't work, then we loop and
try again.  see http://lwn.net/Articles/392897/ for the theory
behind this.  we still have a lot of racy issues wakeup issues,
but this is a first step to closing some of them.