port powerd to XO-3
the first XO-3 doesn't have a DCON, which complicates things
quite a bit.  it does have a brightness control.  separate the DCON
and brightness conditionals, add an emulation of turning the dcon
off (so that brightness changes are still honored while it's
off), and add support for touchscreen as wakeup device.