treat 1.75 wlan wakeups same as unknown
the behavior for each is the same:  treat them as user activity that
doesn't brighten the screen if it's been dimmed or blanked.

this conveniently sidesteps the problem of actually identifying
wlan wakeups -- there's no easy correlation between card interrupts
and "packets that we want to wake us", so the wakeup_counter mechanism
fails us.