rework camera busyness checks for 1.75, and new siv120d camera
the XO-1.75/3 camera is similar to XO-1 (and unlike 1.5), in that
the interrupt remains claimed at all times, so we need to count
interrupts to detect that it's in use.  in addition, the newer
siv120d cameras lower the frame rate in low light, so we need to
set the interrupt threshold much lower to prevent unwanted suspends.