treat some unknown wakeups as RTC
if we get an "unknown" wakeup at just about the time that the RTC
was scheduled to go off, treat it as such.  this is probably the
right thing to do, but during 3.3 x86 development, while the RTC
doesn't report wakeup counts that would identify it, it's the
right hack.