fall back to device tree if olpc-hwinfo is unavailable
olpc-hwinfo is shipped as part of olpc-utils, which does a lot of
changes to the system that are not suitable for most systems other
than XOs running OLPC OS. olpc-hwinfo will still be used if available,
but in addition to reading the DMI information (only available on x86,
i.e. on XO-1 and XO-1.5) we now check the device tree (if available),
thereby adding XO-1.75 support.

[ removed unfamiliar bash-ism -pgf ]

Signed-off-by: Sascha Silbe <silbe@activitycentral.com>