protect against "negative" dates
we toss events prior to $eventcutoff, which is initialized to 0.
if date(1) can be convinced to return a negative number, then it
causes us to toss all events.  this would be fixed at the first
suspend/resume cycle, but without events we'll never get one.
things seem to work if we initialize $eventcutoff to a very
negative number, instead.   fixes #12621